+++ src/api/operations/get-customer-wishlist.ts
@@ -61,7 +61,7 @@ async function getCustomerWishlist({
 
     if (entityIds?.length) {
       const graphqlData = await getAllProducts({
-        variables: { first: 100, entityIds },
+        variables: { first: 50, entityIds },
         config,
       })
       // Put the products in an object that we can use to get them by id